Initial work with http crate
